fix: restore the Equals/GetHashCode contract on B2Vec2
Moving operator== to an IEEE compare was the right call, it matches C.
But Equals just forwarded to the operator, so a NaN vector stopped being
equal to itself and stopped agreeing with GetHashCode. That breaks
Dictionary<B2Vec2, _>. Keep the operator on IEEE semantics and compare
bitwise in Equals.

While here:

- b2World_Draw takes the debug draw by in. C passes b2DebugDraw*, and
  the struct has 27 fields, so it was copied every frame.
- b2GrowBitSet guards its copy the way b2CreateBitSet already guards the
  fill. Growing a bit set created with zero capacity threw, because
  b2Alloc hands back null there.
- b2ClearBit takes an int like the other three, so the call sites drop
  their casts.
- b2InternalAssert is public. It is B2_API upstream.

+++ src/Box2D.NET/B2SolverSets.cs
@@ -330,8 +330,8 @@ namespace Box2D.NET
                     if (colorIndex != B2_OVERFLOW_INDEX)
                     {
                         // might clear a bit for a static body, but this has no effect
-                        b2ClearBit(ref color.bodySet, (uint)contact.edges[0].bodyId);
-                        b2ClearBit(ref color.bodySet, (uint)contact.edges[1].bodyId);
+                        b2ClearBit(ref color.bodySet, contact.edges[0].bodyId);
+                        b2ClearBit(ref color.bodySet, contact.edges[1].bodyId);
                     }
 
                     int localIndex = contact.localIndex;
@@ -380,8 +380,8 @@ namespace Box2D.NET
                     if (colorIndex != B2_OVERFLOW_INDEX)
                     {
                         // might clear a bit for a static body, but this has no effect
-                        b2ClearBit(ref color.bodySet, (uint)joint.edges[0].bodyId);
-                        b2ClearBit(ref color.bodySet, (uint)joint.edges[1].bodyId);
+                        b2ClearBit(ref color.bodySet, joint.edges[0].bodyId);
+                        b2ClearBit(ref color.bodySet, joint.edges[1].bodyId);
                     }
 
                     int sleepJointIndex = sleepSet.jointSims.count;
